+++ packages/core/package.json
@@ -1,19 +1,9 @@
 {
 	"name": "hybrid",
 	"version": "1.3.2",
-	"keywords": [
-		"xmtp",
-		"agent",
-		"ai",
-		"hybrid",
-		"blockchain",
-		"messaging"
-	],
+	"keywords": ["xmtp", "agent", "ai", "hybrid", "blockchain", "messaging"],
 	"type": "module",
-	"files": [
-		"dist",
-		"src"
-	],
+	"files": ["dist", "src"],
 	"exports": {
 		".": {
 			"types": "./dist/index.d.ts",
@@ -43,6 +33,7 @@
 		"@coinbase/coinbase-sdk": "^0.22.0",
 		"@hono/node-server": "^1.13.5",
 		"@hybrd/ponder": "workspace:*",
+		"@hybrd/types": "workspace:*",
 		"@hybrd/utils": "workspace:*",
 		"@hybrd/xmtp": "workspace:*",
 		"ai": "^5.0.28",
