+++ packages/ponder/package.json
@@ -18,6 +18,7 @@
 		"format": "biome format --write"
 	},
 	"dependencies": {
+		"@hybrd/types": "workspace:*",
 		"hono": "catalog:stack"
 	},
 	"devDependencies": {
