+++ packages/types/biome.jsonc
@@ -1,0 +1,3 @@
+{
+	"extends": ["@config/biome"]
+}
