+++ packages/types/package.json
@@ -1,7 +1,9 @@
 {
-	"name": "@hybrd/xmtp",
+	"name": "@hybrd/types",
 	"version": "1.3.2",
+	"keywords": ["xmtp", "agent", "ai", "hybrid", "types", "typescript"],
 	"type": "module",
+	"files": ["dist", "src"],
 	"exports": {
 		".": {
 			"types": "./dist/index.d.ts",
@@ -11,31 +13,31 @@
 	},
 	"scripts": {
 		"build": "tsup",
+		"build:watch": "tsup --watch",
 		"clean": "rm -rf .turbo dist",
+		"test": "vitest",
+		"test:run": "vitest run",
 		"typecheck": "tsc --noEmit",
 		"lint": "biome lint --unsafe",
 		"lint:fix": "biome lint --write --unsafe",
 		"format": "biome format --write"
 	},
 	"dependencies": {
-		"@hybrd/types": "workspace:*",
-		"@xmtp/agent-sdk": "^0.0.12",
-		"@xmtp/content-type-group-updated": "2.0.2",
-		"@xmtp/content-type-reaction": "2.0.2",
-		"@xmtp/content-type-reply": "2.0.2",
-		"@xmtp/content-type-text": "2.0.2",
-		"@xmtp/content-type-transaction-reference": "2.0.2",
-		"@xmtp/content-type-wallet-send-calls": "2.0.0",
-		"@xmtp/node-sdk": "^4.0.1",
-		"hono": "^4.7.4",
-		"jsonwebtoken": "^9.0.2",
-		"uint8arrays": "^5.1.0",
-		"viem": "^2.22.17"
+		"ai": "^5.0.28",
+		"zod": "^3.23.8"
+	},
+	"peerDependencies": {
+		"@xmtp/content-type-group-updated": "*",
+		"@xmtp/content-type-reaction": "*",
+		"@xmtp/content-type-reply": "*",
+		"@xmtp/content-type-transaction-reference": "*",
+		"@xmtp/content-type-wallet-send-calls": "*",
+		"@xmtp/node-sdk": "*",
+		"hono": "*"
 	},
 	"devDependencies": {
 		"@config/biome": "workspace:*",
 		"@config/tsconfig": "workspace:*",
-		"@types/jsonwebtoken": "^9.0.7",
 		"@types/node": "catalog:stack",
 		"tsup": "^8.5.0",
 		"vitest": "^3.2.4"
