+++ packages/types/src/index.ts
@@ -1,0 +1,42 @@
+// Agent types
+export type {
+	AgentConfig,
+	DefaultRuntimeExtension,
+	GenerateOptions,
+	StreamOptions,
+	ToolGenerator
+} from "./agent"
+
+// Tool types
+export type {
+	AnyTool,
+	Tool,
+	ToolConfig
+} from "./tool"
+
+// Plugin types
+export type {
+	Plugin,
+	PluginContext,
+	PluginRegistry
+} from "./plugin"
+
+// Runtime types
+export type {
+	AgentRuntime,
+	BaseRuntime
+} from "./runtime"
+
+// XMTP types
+export type {
+	HonoVariables,
+	XmtpClient,
+	XmtpConversation,
+	XmtpCredentials,
+	XmtpMessage,
+	XmtpSender,
+	XmtpSubjects
+} from "./xmtp"
+
+// Resolver types
+export type { Resolver } from "./resolver"
