+++ packages/types/src/resolver.ts
@@ -1,0 +1,6 @@
+/**
+ * Resolver interface for address resolution
+ */
+export interface Resolver {
+	resolve: (address: string) => Promise<string | null>
+}
