+++ packages/types/src/runtime.ts
@@ -1,0 +1,31 @@
+import { UIMessage } from "ai"
+import type {
+	XmtpClient,
+	XmtpConversation,
+	XmtpMessage,
+	XmtpSender,
+	XmtpSubjects
+} from "./xmtp"
+
+/**
+ * Base runtime that is always included in agents and tools.
+ * Additional fields can be added via generic type parameters as intersections.
+ *
+ * Example usage:
+ * - Basic: AgentRuntime (just chatId and messages)
+ * - Extended: AgentRuntime & { userId: string } (adds userId field)
+ */
+export interface BaseRuntime extends Record<string, unknown> {
+	conversation: XmtpConversation
+	message: XmtpMessage
+	parentMessage?: XmtpMessage
+	rootMessage: XmtpMessage
+	sender: XmtpSender
+	subjects: XmtpSubjects
+	xmtpClient: XmtpClient
+}
+
+export type AgentRuntime = BaseRuntime & {
+	chatId?: string
+	messages: Array<UIMessage>
+}
