+++ packages/types/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "@config/tsconfig/base.json",
+	"compilerOptions": {
+		"outDir": "dist"
+	},
+	"include": ["src/**/*"],
+	"exclude": ["dist", "node_modules"]
+}
